separate sources and object files
