+++ src/equation_solver.cpp
