+++ src/equation_solver.h
